requirements.txt
gooddata-sdk~=1.4.0
gooddata-pandas~=1.4.0
schedule~=1.2.0
boto3~=1.28.27
PyYAML~=6.0.1